# contracts/openapi/tool-defect-api-v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "刀具缺陷检测系统接口",
    "version": "1.0.0",
    "description": "采集、业务查询、复核、数据集、训练、模型发布和内部推理接口的 v1 唯一契约源。"
  },
  "servers": [
    {
      "url": "https://api.example.invalid",
      "description": "地址由环境配置提供，示例域名不可用于生产"
    }
  ],
...
  "paths": {
...
    "/internal/v1/detection-tasks/{detection_task_id}/attempts": {
      "post": {
        "operationId": "startDetectionAttempt",
        "tags": [
          "internal"
        ],
        "security": [
          {
            "ServiceMtls": [],
            "ServiceOAuth": [
              "inference:callback"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DetectionTaskId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttemptStartRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "执行尝试已创建",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptStartResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "startDetectionAttempt"
      }
    },
    "/internal/v1/detection-attempts/{attempt_id}/result": {
      "put": {
        "operationId": "submitDetectionResult",
        "tags": [
          "internal"
        ],
        "security": [
          {
            "ServiceMtls": [],
            "ServiceOAuth": [
              "inference:callback"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AttemptId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../json-schema/detection-result-v1.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "结果已幂等接受",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "submitDetectionResult"
      }
    },
    "/internal/v1/detection-attempts/{attempt_id}/failure": {
      "put": {
        "operationId": "submitDetectionFailure",
        "tags": [
          "internal"
        ],
        "security": [
          {
            "ServiceMtls": [],
            "ServiceOAuth": [
              "inference:callback"
            ]
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AttemptId"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/Traceparent"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectionFailureRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "失败事实已接受，重试决策由后端负责",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acknowledgement"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "x-write-example-operation": "submitDetectionFailure"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "EdgeDeviceMtls": {
        "type": "mutualTLS"
      },
      "EdgeOAuth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://identity.example.invalid/oauth2/token",
            "scopes": {
              "capture:write": "写入采集事件",
              "capture:read": "读取采集状态",
              "device:heartbeat": "上报设备心跳"
            }
          }
        }
      },
      "UserOidc": {
        "type": "openIdConnect",
        "openIdConnectUrl": "https://identity.example.invalid/.well-known/openid-configuration"
      },
      "ServiceMtls": {
        "type": "mutualTLS"
      },
      "ServiceOAuth": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://identity.example.invalid/oauth2/token",
            "scopes": {
              "inference:callback": "提交推理执行结果",
              "runtime:read": "读取推理运行时健康"
            }
          }
        }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 256
        }
      },
...
      "Traceparent": {
        "name": "traceparent",
        "in": "header",
        "required": false,
        "schema": {
          "$ref": "../json-schema/common-v1.schema.json#/$defs/Traceparent"
        }
      },
...
      "DetectionTaskId": {
        "name": "detection_task_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
        }
      },
      "AttemptId": {
        "name": "attempt_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
        }
      },
...
    },
...
    "responses": {
      "BadRequest": {
        "description": "请求格式错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "身份认证失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "Forbidden": {
        "description": "没有资源或操作权限",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "幂等摘要或乐观锁冲突",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
      "ValidationError": {
        "description": "语义或状态校验失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../json-schema/standard-error-v1.schema.json"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "AttemptStartRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message_id",
          "worker_id",
          "runtime_version",
          "model_sha256"
        ],
        "properties": {
          "message_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "worker_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "runtime_version": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Version"
          },
          "model_sha256": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Sha256"
          }
        }
      },
      "AttemptStartResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "attempt_id",
          "attempt_no",
          "status"
        ],
        "properties": {
          "attempt_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          },
          "attempt_no": {
            "type": "integer",
            "minimum": 1
          },
          "status": {
            "const": "RUNNING"
          }
        }
      },
      "DetectionFailureRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error_code",
          "stage",
          "retryable",
          "message",
          "occurred_at"
        ],
        "properties": {
          "error_code": {
            "type": "string",
            "pattern": "^TD-[A-Z]+-[A-Z]+-[0-9]{3}$"
          },
          "stage": {
            "type": "string",
            "enum": [
              "DOWNLOAD",
              "DECODE",
              "PREPROCESS",
              "MODEL_LOAD",
              "INFERENCE",
              "POSTPROCESS",
              "UPLOAD",
              "CALLBACK"
            ]
          },
          "retryable": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "occurred_at": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/UtcTimestamp"
          }
        }
      },
      "ResultAcceptedResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accepted",
          "result_sha256"
        ],
        "properties": {
          "accepted": {
            "const": true
          },
          "result_sha256": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Sha256"
          }
        }
      },
      "Acknowledgement": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accepted",
          "request_id"
        ],
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "request_id": {
            "$ref": "../json-schema/common-v1.schema.json#/$defs/Uuid"
          }
        }
      },
...
    }
  },
...
}